+++ pyrehau_neasmart/exceptions.py
@@ -1,0 +1,8 @@
+class RehauNeaSmartError(Exception):
+
+    def __init__(self, value):
+        self.value = value
+
+    # __str__ is to print() the value
+    def __str__(self):
+        return(repr(self.value))